Separate hashing, tree-building, and utilities from algorithm.go
