+++ src/libraries/filcrypto/filproofs/algorithms.go
@@ -62,6 +62,7 @@ func PoStCfg(pType sector.PoStType, sectorSize sector.SectorSize, partitions UIn
 		ProofInstance_: sector.PROOFS[UInt(registeredProof)],
 	}
 }
+
 func SurprisePoStCfg(sectorSize sector.SectorSize) *sector.PoStCfg_I {
 	return PoStCfg(sector.PoStType_SurprisePoSt, sectorSize, SURPRISE_POST_PARTITIONS)
 }
@@ -79,98 +80,6 @@ func ElectionPoStVerifier(cfg sector.PoStCfg) *PoStVerifier_I {
 func SurprisePoStVerifier(cfg sector.PoStCfg) *PoStVerifier_I {
 	return &PoStVerifier_I{
 		SurprisePoStCfg_: cfg,
-	}
-}
-
-func WinSDRParams(cfg ProofsCfg) *WinStackedDRG_I {
-	// TODO: Bridge constants with orient model.
-	const LAYERS = 10
-	const OFFLINE_CHALLENGES = 6666
-	const OFFLINE_WINDOW_CHALLENGES = 1111
-	const FEISTEL_ROUNDS = 3
-	var FEISTEL_KEYS = [FEISTEL_ROUNDS]UInt{1, 2, 3}
-	var FIELD_MODULUS = new(big.Int)
-	// https://github.com/zkcrypto/pairing/blob/master/src/bls12_381/fr.rs#L4
-	FIELD_MODULUS.SetString("52435875175126190479447740508185965837690552500527637822603658699938581184513", 10)
-
-	nodes := UInt(cfg.SealCfg().SectorSize() / NODE_SIZE)
-
-	return &WinStackedDRG_I{
-		Layers_:           WinStackedDRGLayers(LAYERS),
-		Challenges_:       WinStackedDRGChallenges(OFFLINE_CHALLENGES),
-		WindowChallenges_: WinStackedDRGWindowChallenges(OFFLINE_WINDOW_CHALLENGES),
-		NodeSize_:         WinStackedDRGNodeSize(NODE_SIZE),
-		Nodes_:            WinStackedDRGNodes(nodes),
-		Algorithm_:        &WinStackedDRG_Algorithm_I{},
-		DRGCfg_: &DRGCfg_I{
-			Algorithm_: &DRGCfg_Algorithm_I{
-				ParentsAlgorithm_: DRGCfg_Algorithm_ParentsAlgorithm_DRSample,
-				RNGAlgorithm_:     DRGCfg_Algorithm_RNGAlgorithm_ChaCha20,
-			},
-			Degree_: 6,
-			Nodes_:  DRGNodeCount(nodes),
-		},
-		ExpanderGraphCfg_: &ExpanderGraphCfg_I{
-			Algorithm_: ExpanderGraphCfg_Algorithm_Make_ChungExpanderAlgorithm(
-				&ChungExpanderAlgorithm_I{
-					PermutationAlgorithm_: ChungExpanderAlgorithm_PermutationAlgorithm_Make_Feistel(&Feistel_I{
-						Keys_:         FEISTEL_KEYS[:],
-						Rounds_:       FEISTEL_ROUNDS,
-						HashFunction_: ChungExpanderPermutationFeistelHashFunction_SHA256,
-					}),
-				}),
-			Degree_: 8,
-			Nodes_:  ExpanderGraphNodeCount(nodes),
-		},
-		WindowDRGCfg_: &DRGCfg_I{
-			Algorithm_: &DRGCfg_Algorithm_I{
-				ParentsAlgorithm_: DRGCfg_Algorithm_ParentsAlgorithm_DRSample,
-				RNGAlgorithm_:     DRGCfg_Algorithm_RNGAlgorithm_ChaCha20,
-			},
-			Degree_: 0,
-			Nodes_:  DRGNodeCount(nodes),
-		},
-		WindowExpanderGraphCfg_: &ExpanderGraphCfg_I{
-			Algorithm_: ExpanderGraphCfg_Algorithm_Make_ChungExpanderAlgorithm(
-				&ChungExpanderAlgorithm_I{
-					PermutationAlgorithm_: ChungExpanderAlgorithm_PermutationAlgorithm_Make_Feistel(&Feistel_I{
-						Keys_:         FEISTEL_KEYS[:],
-						Rounds_:       FEISTEL_ROUNDS,
-						HashFunction_: ChungExpanderPermutationFeistelHashFunction_SHA256,
-					}),
-				}),
-			Degree_: 8,
-			Nodes_:  ExpanderGraphNodeCount(nodes),
-		},
-
-		Curve_: &EllipticCurve_I{
-			FieldModulus_: *FIELD_MODULUS,
-		},
-		Cfg_: cfg,
-	}
-}
-
-func (sdr *WinStackedDRG_I) Drg() *DRG_I {
-	return &DRG_I{
-		Config_: sdr.DRGCfg(),
-	}
-}
-
-func (sdr *WinStackedDRG_I) Expander() *ExpanderGraph_I {
-	return &ExpanderGraph_I{
-		Config_: sdr.ExpanderGraphCfg(),
-	}
-}
-
-func (sdr *WinStackedDRG_I) WindowDrg() *DRG_I {
-	return &DRG_I{
-		Config_: sdr.WindowDRGCfg(),
-	}
-}
-
-func (sdr *WinStackedDRG_I) WindowExpander() *ExpanderGraph_I {
-	return &ExpanderGraph_I{
-		Config_: sdr.WindowExpanderGraphCfg(),
 	}
 }
 
@@ -255,77 +164,6 @@ func (chung *ChungExpanderAlgorithm_I) _ithParent(node UInt, i UInt, degree Expa
 func (f *Feistel_I) Permute(size UInt, i UInt) UInt {
 	// Call into feistel.go.
 	panic("TODO")
-}
-
-func (sdr *WinStackedDRG_I) Seal(sid sector.SectorID, data []byte, randomness sector.SealRandomness) SealSetupArtifacts {
-	windowCount := int(sdr.WindowCount())
-	nodeSize := int(sdr.NodeSize())
-	nodes := int(sdr.Nodes())
-	curveModulus := sdr.Curve().FieldModulus()
-
-	var windowData [][]byte
-
-	for i := 0; i < len(data); i += nodeSize {
-		windowData = append(windowData, data[i*nodeSize:(i+1)*nodeSize])
-	}
-
-	util.Assert(len(windowData) == windowCount)
-
-	var windowKeyLayers [][]byte
-	var finalWindowKeyLayer []byte
-
-	commD, commDTreePath := ComputeDataCommitment(data)
-	sealSeed := computeSealSeed(sid, randomness, commD)
-
-	for i := 0; i < windowCount; i++ {
-		keyLayers := sdr._generateWindowKey(sealSeed, i, sid, commD, nodes, randomness)
-
-		lastIndex := len(keyLayers) - 1
-		windowKeyLayers = append(windowKeyLayers, keyLayers[:lastIndex]...)
-		finalWindowKeyLayer = append(finalWindowKeyLayer, keyLayers[lastIndex]...)
-	}
-
-	qLayer := encodeDataInPlace(data, finalWindowKeyLayer, nodeSize, &curveModulus)
-	// NOTE: qLayer and data are now the same, and qLayer is introduced here for descriptive clarity only.
-
-	replica := labelLayer(sdr.Drg(), sdr.Expander(), sealSeed, WRAPPER_LAYER_WINDOW_INDEX, nodes, nodeSize, qLayer)
-
-	commC, commQ, commRLast, commR, commCTreePath, commQTreePath, commRLastTreePath := sdr.GenerateCommitments(replica, windowKeyLayers, qLayer)
-
-	result := SealSetupArtifacts_I{
-		CommD_:             Commitment(commD),
-		CommR_:             SealedSectorCID(commR),
-		CommC_:             Commitment(commC),
-		CommQ_:             Commitment(commQ),
-		CommRLast_:         Commitment(commRLast),
-		CommDTreePath_:     commDTreePath,
-		CommCTreePath_:     commCTreePath,
-		CommQTreePath_:     commQTreePath,
-		CommRLastTreePath_: commRLastTreePath,
-		Seed_:              sealSeed,
-		KeyLayers_:         windowKeyLayers,
-		Replica_:           replica,
-	}
-	return &result
-}
-
-func (sdr *WinStackedDRG_I) _generateWindowKey(sealSeed sector.SealSeed, windowIndex int, sid sector.SectorID, commD sector.UnsealedSectorCID, nodes int, randomness sector.SealRandomness) [][]byte {
-	nodeSize := int(sdr.NodeSize())
-	curveModulus := sdr.Curve().FieldModulus()
-	layers := int(sdr.Layers())
-
-	keyLayers := generateSDRKeyLayers(sdr.WindowDrg(), sdr.WindowExpander(), sealSeed, windowIndex, nodes, layers, nodeSize, curveModulus)
-
-	return keyLayers
-}
-
-func (sdr *WinStackedDRG_I) GenerateCommitments(replica []byte, windowKeyLayers [][]byte, qLayer []byte) (commC PedersenHash, commQ PedersenHash, commRLast PedersenHash, commR PedersenHash, commCTreePath file.Path, commQTreePath file.Path, commRLastTreePath file.Path) {
-	commC, commCTreePath = computeCommC(windowKeyLayers, int(sdr.NodeSize()))
-	commQ, commQTreePath = computeCommQ(qLayer, int(sdr.NodeSize()))
-	commRLast, commRLastTreePath = BuildTree_PedersenHash(replica)
-	commR = TernaryHash_PedersenHash(commC, commQ, commRLast)
-
-	return commC, commQ, commRLast, commR, commCTreePath, commQTreePath, commRLastTreePath
 }
 
 func getProverID(minerID addr.ActorID) []byte {
@@ -467,169 +305,6 @@ func hashColumn(column []Label) PedersenHash {
 	return HashBytes_PedersenHash(preimage)
 }
 
-func (sdr *WinStackedDRG_I) CreateSealProof(challengeSeed sector.InteractiveSealRandomness, aux sector.ProofAuxTmp) sector.SealProof {
-	privateProof := sdr.CreatePrivateSealProof(challengeSeed, aux)
-
-	// Sanity check: newly-created proofs must pass verification.
-	util.Assert(sdr.VerifyPrivateSealProof(privateProof, aux.Seed(), challengeSeed, aux.CommD(), aux.CommR()))
-
-	return sdr.CreateOfflineCircuitProof(privateProof, aux)
-}
-
-func (sdr *WinStackedDRG_I) CreatePrivateSealProof(randomness sector.InteractiveSealRandomness, aux sector.ProofAuxTmp) (privateProof PrivateOfflineProof) {
-	sealSeed := aux.Seed()
-	nodeSize := UInt(sdr.NodeSize())
-	wrapperChallenges, windowChallenges := sdr._generateOfflineChallenges(sealSeed, randomness, sdr.Challenges(), sdr.WindowChallenges())
-
-	dataTree := LoadMerkleTree(aux.CommDTreePath())
-	columnTree := LoadMerkleTree(aux.CommCTreePath())
-	replicaTree := LoadMerkleTree(aux.PersistentAux().CommRLastTreePath())
-	qTree := LoadMerkleTree(aux.CommQTreePath())
-
-	windows := int(sdr.WindowCount())
-	windowSize := int(uint64(sdr.Cfg().SealCfg().SectorSize()) / UInt(sdr.WindowCount()))
-
-	for c := range windowChallenges {
-		columnProofs := createColumnProofs(sdr.WindowDrg(), sdr.WindowExpander(), UInt(c), nodeSize, columnTree, aux, windows, windowSize)
-		privateProof.ColumnProofs = append(privateProof.ColumnProofs, columnProofs...)
-
-		windowProof := createWindowProof(sdr.WindowDrg(), sdr.WindowExpander(), UInt(c), nodeSize, dataTree, columnTree, qTree, aux, windows, windowSize)
-		privateProof.WindowProofs = append(privateProof.WindowProofs, windowProof)
-	}
-
-	for c := range wrapperChallenges {
-		wrapperProof := createWrapperProof(sdr.Drg(), sdr.Expander(), sealSeed, UInt(c), nodeSize, qTree, replicaTree, aux, windows, windowSize)
-		privateProof.WrapperProofs = append(privateProof.WrapperProofs, wrapperProof)
-	}
-
-	return privateProof
-}
-
-// Verify a private proof.
-// NOTE: Verification of a private proof is exactly the computation we will prove we have performed in a zk-SNARK.
-// If we can verifiably prove that we have performed the verification of a private proof, then we need not reveal the proof itself.
-// Since the zk-SNARK circuit proof is much smaller than the private proof, this allows us to save space on the chain (at the cost of increased computation to generate the zk-SNARK proof).
-func (sdr *WinStackedDRG_I) VerifyPrivateSealProof(privateProof PrivateOfflineProof, sealSeed sector.SealSeed, randomness sector.InteractiveSealRandomness, commD Commitment, commR sector.SealedSectorCID) bool {
-	nodeSize := int(sdr.NodeSize())
-	windowCount := int(sdr.WindowCount())
-	windowSize := int(UInt(sdr.Cfg().SealCfg().SectorSize()) / UInt(sdr.WindowCount())) // TOOD: Make this a function.
-	layers := int(sdr.Layers())
-	curveModulus := sdr.Curve().FieldModulus()
-	windowChallenges, wrapperChallenges := sdr._generateOfflineChallenges(sealSeed, randomness, sdr.Challenges(), sdr.WindowChallenges())
-
-	windowProofs := privateProof.WindowProofs
-	columnProofs := privateProof.ColumnProofs
-	wrapperProofs := privateProof.WrapperProofs
-
-	// commC, commQ, and commRLast must be the same for all challenge proofs, so we can arbitrarily verify against the first.
-	firstColumnProof := columnProofs[0]
-	firstWrapperProof := wrapperProofs[0]
-	commC := firstColumnProof.InclusionProof.Root()
-	commQ := firstWrapperProof.QLayerProofs[0].Root()
-	commRLast := firstWrapperProof.ReplicaProof.Root()
-
-	windowDrgParentCount := int(sdr.WindowDRGCfg().Degree())
-	windowExpanderParentCount := int(sdr.WindowDRGCfg().Degree())
-	wrapperExpanderParentCount := int(sdr.ExpanderGraphCfg().Degree())
-
-	for i, challenge := range windowChallenges {
-		// Verify one OfflineSDRChallengeProof.
-		windowProof := windowProofs[i]
-		dataProof := windowProof.DataProof
-		qLayerProof := windowProof.QLayerProof
-
-		// Verify column proofs and that they represent the right columns.
-		columnElements := getColumnElements(sdr.Drg(), sdr.Expander(), challenge)
-
-		// Check column openings.
-		for i, columnElement := range columnElements {
-			columnProof := columnProofs[i]
-
-			// The provided column proofs must correspond to the expected columns.
-			if !columnProof.Verify(commC, UInt(columnElement)) {
-				return false
-			}
-		}
-
-		// Check labeling.
-		for w := 0; w < windowCount; w++ {
-			for layer := 0; layer < layers; layer++ {
-				var parents []Label
-
-				// First column proof is the challenge.
-				// Then the DRG parents.
-				for _, drgParentProof := range columnProofs[1 : 1+windowDrgParentCount] {
-					parent := drgParentProof.Column[layer]
-					parents = append(parents, parent)
-				}
-				// And the expander parents, if not the first layer.
-				if layer > 0 {
-					for _, expanderParentProof := range columnProofs[1+windowDrgParentCount : 1+windowExpanderParentCount] {
-						parent := expanderParentProof.Column[layer-1]
-						parents = append(parents, parent)
-					}
-				}
-
-				calculatedLabel := generateLabel(sealSeed, i, w, parents)
-
-				if layer == layers-1 {
-					// Last layer includes encoding.
-					dataNode := dataProof.Leaf()
-					qLayerNode := qLayerProof.Leaf()
-
-					if !dataProof.Verify(commD, UInt(windowSize*w)+challenge) {
-						return false
-					}
-
-					encodedNode := encodeNode(dataNode, calculatedLabel, &curveModulus, nodeSize)
-
-					if !bytes.Equal(encodedNode, qLayerNode) {
-						return false
-					}
-
-				} else {
-					providedLabel := columnProofs[columnElements[0]].Column[layer]
-
-					if !bytes.Equal(calculatedLabel, providedLabel) {
-						return false
-					}
-				}
-			}
-		}
-	}
-
-	for i, challenge := range wrapperChallenges {
-		wrapperProof := wrapperProofs[i]
-		replicaProof := wrapperProof.ReplicaProof
-		qLayerProofs := wrapperProof.QLayerProofs
-
-		if !replicaProof.Verify(commRLast, challenge) {
-			return false
-		}
-
-		var parents []Label
-		for i := 0; i < wrapperExpanderParentCount; i++ {
-			parent := qLayerProofs[i].Leaf()
-			parents = append(parents, parent)
-		}
-
-		label := generateLabel(sealSeed, i, windowCount+1, parents)
-		replicaNode := replicaProof.Leaf()
-
-		if !bytes.Equal(label, replicaNode) {
-			return false
-		}
-	}
-
-	commRCalculated := TernaryHash_PedersenHash(commC, commQ, commRLast)
-
-	if !bytes.Equal(commRCalculated, AsBytes_SealedSectorCID(commR)) {
-		return false
-	}
-
-	return true
-}
-
 func createColumnProofs(drg *DRG_I, expander *ExpanderGraph_I, challenge UInt, nodeSize UInt, columnTree MerkleTree, aux sector.ProofAuxTmp, windows int, windowSize int) []SDRColumnProof {
 	columnElements := getColumnElements(drg, expander, challenge)
 
@@ -767,36 +442,6 @@ func (proof *SDRColumnProof) Verify(root []byte, challenge UInt) bool {
 	return proof.InclusionProof.Verify(root, challenge)
 }
 
-func (sdr *WinStackedDRG_I) CreateOfflineCircuitProof(proof PrivateOfflineProof, aux sector.ProofAuxTmp) sector.SealProof {
-	// partitions := sdr.Partitions()
-	// publicInputs := GeneratePublicInputs()
-
-	panic("TODO")
-	var bytes []byte
-
-	var registeredProof sector.RegisteredProof
-
-	switch sdr.Cfg().SealCfg().SectorSize() {
-	case GIB_32:
-		registeredProof = sector.RegisteredProof_WinStackedDRG32GiBSeal
-	}
-
-	sealProof := sector.SealProof_I{
-		// TODO: This must also depend on the sector size if more than one size is in the spec.
-		ProofInstance_: sector.PROOFS[UInt(registeredProof)],
-		ProofBytes_:    bytes,
-	}
-
-	return &sealProof
-}
-
-func (sdr *WinStackedDRG_I) _generateOfflineChallenges(sealSeed sector.SealSeed, randomness sector.InteractiveSealRandomness, wrapperChallengeCount WinStackedDRGChallenges, windowChallengeCount WinStackedDRGWindowChallenges) (windowChallenges []UInt, wrapperChallenges []UInt) {
-	wrapperChallenges = generateOfflineChallenges(int(sdr.Nodes()), sealSeed, randomness, int(wrapperChallengeCount))
-	windowChallenges = generateOfflineChallenges(int(sdr.WindowDRGCfg().Nodes()), sealSeed, randomness, int(windowChallengeCount))
-
-	return windowChallenges, wrapperChallenges
-}
-
 func generateOfflineChallenges(challengeRange int, sealSeed sector.SealSeed, randomness sector.InteractiveSealRandomness, challengeCount int) []UInt {
 	var challenges []UInt
 	challengeRangeSize := challengeRange - 1 // Never challenge the first node.
@@ -841,17 +486,6 @@ func addEncode(data []byte, key []byte, modulus *big.Int, nodeSize int) []byte {
 
 ////////////////////////////////////////////////////////////////////////////////
 // Seal Verification
-
-func (sdr *WinStackedDRG_I) VerifySeal(sv sector.SealVerifyInfo) bool {
-	onChain := sv.OnChain()
-	sealProof := onChain.Proof()
-	commR := sector.SealedSectorCID(onChain.SealedCID())
-	commD := sector.UnsealedSectorCID(sv.UnsealedCID())
-	sealSeed := computeSealSeed(sv.SectorID(), sv.Randomness(), commD)
-
-	wrapperChallenges, windowChallenges := sdr._generateOfflineChallenges(sealSeed, sv.InteractiveRandomness(), sdr.Challenges(), sdr.WindowChallenges())
-	return sdr._verifyOfflineCircuitProof(commD, commR, sealSeed, windowChallenges, wrapperChallenges, sealProof)
-}
 
 func ComputeUnsealedSectorCIDFromPieceInfos(sectorSize sector.SectorSize, pieceInfos []PieceInfo) (unsealedCID sector.UnsealedSectorCID, err error) {
 	rootPieceInfo := computeRootPieceInfo(pieceInfos)
@@ -944,11 +578,6 @@ func joinPieceInfos(left PieceInfo, right PieceInfo) PieceInfo {
 	}
 }
 
-func (sdr *WinStackedDRG_I) _verifyOfflineCircuitProof(commD sector.UnsealedSectorCID, commR sector.SealedSectorCID, sealSeed sector.SealSeed, windowChallenges []UInt, wrapperChallenges []UInt, sv sector.SealProof) bool {
-	//publicInputs := GeneratePublicInputs()
-	panic("TODO")
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 // PoSt
 
@@ -998,43 +627,6 @@ func generateCandidate(algorithm sector.ProofAlgorithm, cfg sector.PoStCfg, rand
 		candidate = sdr._generateCandidate(cfg, randomness, aux, sectorID, sectorChallengeIndex)
 	}
 	return candidate
-}
-
-func (sdr *WinStackedDRG_I) _generateCandidate(postCfg sector.PoStCfg, randomness sector.PoStRandomness, aux sector.PersistentProofAux, sectorID sector.SectorID, sectorChallengeIndex UInt) sector.PoStCandidate {
-	cfg := postCfg.InstanceCfg().As_PoStCfgV1()
-
-	nodes := int(cfg.Nodes())
-	leafChallengeCount := int(cfg.LeafChallengeCount())
-	challengeRangeSize := int(cfg.ChallengeRangeSize())
-	treePath := aux.CommRLastTreePath()
-	tree := LoadMerkleTree(treePath)
-
-	var data []byte
-	var inclusionProofs []InclusionProof
-	for i := 0; i < leafChallengeCount; i++ {
-		leafChallenge := generateLeafChallenge(randomness, sectorChallengeIndex, i, nodes, challengeRangeSize)
-
-		for j := 0; j < challengeRangeSize; j++ {
-			leafIndex := leafChallenge + UInt(j)
-			data = append(data, tree.Leaf(leafIndex)...)
-			inclusionProof := tree.ProveInclusion(leafIndex)
-			inclusionProofs = append(inclusionProofs, inclusionProof)
-		}
-	}
-
-	partialTicket := computePartialTicket(randomness, sectorID, data)
-
-	privateProof := InternalPrivateCandidateProof{
-		InclusionProofs: inclusionProofs,
-	}
-
-	candidate := sector.PoStCandidate_I{
-		PartialTicket_:  partialTicket,
-		PrivateProof_:   privateProof.externalize(sector.ProofAlgorithm_WinStackedDRGSeal),
-		SectorID_:       sectorID,
-		ChallengeIndex_: sectorChallengeIndex,
-	}
-	return &candidate
 }
 
 func computePartialTicket(randomness sector.PoStRandomness, sectorID sector.SectorID, data []byte) sector.PartialTicket {
@@ -1102,91 +694,6 @@ func newInternalPrivateProof(externalPrivateProof PrivatePostCandidateProof) Int
 	return InternalPrivateCandidateProof{}
 }
 
-func (sdr *WinStackedDRG_I) VerifyInternalPrivateCandidateProof(postCfg sector.PoStCfg, p *InternalPrivateCandidateProof, challengeSeed sector.PoStRandomness, candidate sector.PoStCandidate, commRLast Commitment) bool {
-	cfg := postCfg.InstanceCfg().As_PoStCfgV1()
-	util.Assert(candidate.PrivateProof() == nil)
-	nodes := int(cfg.Nodes())
-	challengeRangeSize := int(cfg.ChallengeRangeSize())
-
-	sectorID := candidate.SectorID()
-	claimedPartialTicket := candidate.PartialTicket()
-
-	allInclusionProofs := p.InclusionProofs
-
-	var ticketData []byte
-
-	for _, p := range allInclusionProofs {
-		ticketData = append(ticketData, p.Leaf()...)
-	}
-
-	// Check partial ticket
-	calculatedTicket := computePartialTicket(challengeSeed, sectorID, ticketData)
-
-	if len(calculatedTicket) != len(claimedPartialTicket) {
-		return false
-	}
-	for i, byte := range claimedPartialTicket {
-		if byte != calculatedTicket[i] {
-			return false
-		}
-	}
-
-	// Helper to get InclusionProofs sequentially.
-	next := func() InclusionProof {
-		if len(allInclusionProofs) < 1 {
-			return nil
-		}
-
-		proof := allInclusionProofs[0]
-		allInclusionProofs = allInclusionProofs[1:]
-		return proof
-	}
-
-	// Check all inclusion proofs.
-	for i := 0; i < int(cfg.LeafChallengeCount()); i++ {
-		leafChallenge := generateLeafChallenge(challengeSeed, candidate.ChallengeIndex(), i, nodes, challengeRangeSize)
-		for j := 0; j < challengeRangeSize; j++ {
-			leafIndex := leafChallenge + UInt(j)
-			proof := next()
-			if proof == nil {
-				// All required inclusion proofs must be provided.
-				return false
-			}
-			if !proof.Verify(commRLast, leafIndex) {
-				return false
-			}
-		}
-	}
-
-	return true
-}
-
-func (sdr *WinStackedDRG_I) VerifyPrivatePoStProof(cfg sector.PoStCfg, privateProof PrivatePoStProof, candidates []sector.PoStCandidate, sectorIDs []sector.SectorID, sectorCommitments sector.SectorCommitments) bool {
-	// This is safe by construction.
-	challengeSeed := privateProof.ChallengeSeed
-
-	for i, p := range privateProof.CandidateProofs {
-		proof := newInternalPrivateProof(p)
-
-		candidate := candidates[i]
-		ci := candidate.ChallengeIndex()
-		expectedSectorID := sectorIDs[ci]
-
-		challengedSectorID := generateSectorChallenge(challengeSeed, i, sectorIDs)
-
-		if expectedSectorID != challengedSectorID {
-			return false
-		}
-
-		commRLast := sectorCommitments[expectedSectorID]
-
-		if !sdr.VerifyInternalPrivateCandidateProof(cfg, &proof, challengeSeed, candidate, commRLast) {
-			return false
-		}
-	}
-	return true
-}
-
 func createPoStCircuitProof(postCfg sector.PoStCfg, algorithm sector.ProofAlgorithm, privateProof PrivatePoStProof) (proof sector.PoStProof) {
 	switch algorithm {
 	case sector.ProofAlgorithm_WinStackedDRGSeal:
@@ -1194,19 +701,6 @@ func createPoStCircuitProof(postCfg sector.PoStCfg, algorithm sector.ProofAlgori
 		proof = sdr._createPoStCircuitProof(postCfg, privateProof)
 	}
 	return proof
-}
-
-func (sdr *WinStackedDRG_I) _createPoStCircuitProof(postCfg sector.PoStCfg, privateProof PrivatePoStProof) sector.PoStProof {
-	panic("TODO")
-
-	cfg := postCfg.InstanceCfg().As_PoStCfgV1()
-
-	postProof := sector.PoStProof_I{
-		Type_:        cfg.Type(),
-		CircuitType_: postCfg.ProofInstance().CircuitType(),
-	}
-
-	return &postProof
 }
 
 func (pv *PoStVerifier_I) _verifyPoStProof(sv sector.PoStVerifyInfo) bool {
@@ -1265,251 +759,4 @@ func CreateSurprisePoStProof(cfg sector.PoStCfg, privateCandidateProofs []Privat
 
 func (pv *PoStVerifier_I) VerifySurprisePoSt(sv sector.PoStVerifyInfo) bool {
 	return pv._verifyPoStProof(sv)
-}
-
-////////////////////////////////////////////////////////////////////////////////
-/// Generic Hashing and Merkle Tree generation
-
-/// Binary hash compression.
-// BinaryHash<T>
-func BinaryHash_T(left []byte, right []byte) util.T {
-	var preimage = append(left, right...)
-	return HashBytes_T(preimage)
-}
-
-func TernaryHash_T(a []byte, b []byte, c []byte) util.T {
-	var preimage = append(a, append(b, c...)...)
-	return HashBytes_T(preimage)
-}
-
-// BinaryHash<PedersenHash>
-func BinaryHash_PedersenHash(left []byte, right []byte) PedersenHash {
-	return PedersenHash{}
-}
-
-func TernaryHash_PedersenHash(a []byte, b []byte, c []byte) PedersenHash {
-	return PedersenHash{}
-}
-
-// BinaryHash<SHA256Hash>
-func BinaryHash_SHA256Hash(left []byte, right []byte) SHA256Hash {
-	result := SHA256Hash{}
-	return trimToFr32(result)
-}
-
-func TernaryHash_SHA256Hash(a []byte, b []byte, c []byte) SHA256Hash {
-	return SHA256Hash{}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-/// Digest
-// HashBytes<T>
-func HashBytes_T(data []byte) util.T {
-	return util.T{}
-}
-
-// HashBytes<PedersenHash>
-func HashBytes_PedersenHash(data []byte) PedersenHash {
-	return PedersenHash{}
-}
-
-// HashBytes<SHA256Hash.
-func HashBytes_SHA256Hash(data []byte) SHA256Hash {
-	// Digest is truncated to 254 bits.
-	result := SHA256Hash{}
-
-	return result
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-func DigestSize_T() int {
-	panic("Unspecialized")
-}
-
-func DigestSize_PedersenHash() int {
-	return 32
-}
-
-func DigestSize_SHA256Hash() int {
-	return 32
-}
-
-////////////////////////////////////////////////////////////////////////////////
-/// Binary Merkle-tree generation
-
-// BuildTree<T>
-func BuildTree_T(data []byte) (util.T, file.Path) {
-	// Plan: define this in terms of BinaryHash_T, then copy-paste changes into T-specific specializations, for now.
-
-	// Nodes are always the digest size so data cannot be compressed to digest for storage.
-	nodeSize := DigestSize_T()
-
-	// TODO: Fail if len(dat) is not a power of 2 and a multiple of the node size.
-
-	rows := [][]byte{data}
-
-	for row := []byte{}; len(row) > nodeSize; {
-		for i := 0; i < len(data); i += 2 * nodeSize {
-			left := data[i : i+nodeSize]
-			right := data[i+nodeSize : i+2*nodeSize]
-
-			hashed := BinaryHash_T(left, right)
-
-			row = append(row, AsBytes_T(hashed)...)
-		}
-		rows = append(rows, row)
-	}
-
-	// Last row is the root
-	root := rows[len(rows)-1]
-
-	if len(root) != nodeSize {
-		panic("math failed us")
-	}
-
-	var filePath file.Path // TODO: dump tree to file.
-	// NOTE: merkle tree file layout is illustrative, not prescriptive.
-
-	// TODO: Check above more carefully. It's just an untested sketch for the moment.
-	return fromBytes_T(root), filePath
-}
-
-// BuildTree<PedersenHash>
-func BuildTree_PedersenHash(data []byte) (PedersenHash, file.Path) {
-	return PedersenHash{}, file.Path("") // FIXME
-}
-
-//  BuildTree<SHA256Hash>
-func BuildTree_SHA256Hash(data []byte) (SHA256Hash, file.Path) {
-	return []byte{}, file.Path("") // FIXME
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-// Destructively trim data so most significant two bits of last byte are 0.
-// This ensure data interpreted as little-endian will not exceed a field with 254-bit capacity.
-// NOTE: 254 bits is the capacity of BLS12-381, but other curves with ~32-byte field elements
-// may have a different capacity. (Example: BLS12-377 has a capacity of 252 bits.)
-func trimToFr32(data []byte) []byte {
-	util.Assert(len(data) == 32)
-	data[31] &= 0x3f // 0x3f = 0b0011_1111
-	return data
-}
-
-func UnsealedSectorCID(h SHA256Hash) sector.UnsealedSectorCID {
-	panic("not implemented -- re-arrange bits")
-}
-
-func SealedSectorCID(h PedersenHash) sector.SealedSectorCID {
-	panic("not implemented -- re-arrange bits")
-}
-
-func Commitment_UnsealedSectorCID(cid sector.UnsealedSectorCID) Commitment {
-	panic("not implemented -- re-arrange bits")
-}
-
-func Commitment_SealedSectorCID(cid sector.SealedSectorCID) Commitment {
-	panic("not implemented -- re-arrange bits")
-}
-
-func ComputeDataCommitment(data []byte) (sector.UnsealedSectorCID, file.Path) {
-	// TODO: make hash parameterizable
-	hash, path := BuildTree_SHA256Hash(data)
-	return UnsealedSectorCID(hash), path
-}
-
-// Compute CommP or CommD.
-func ComputeUnsealedSectorCID(data []byte) (sector.UnsealedSectorCID, file.Path) {
-	// TODO: check that len(data) > minimum piece size and is a power of 2.
-	hash, treePath := BuildTree_SHA256Hash(data)
-	return UnsealedSectorCID(hash), treePath
-}
-
-// Utilities
-
-func reverse(bytes []byte) {
-	for i, j := 0, len(bytes)-1; i < j; i, j = i+1, j-1 {
-		bytes[i], bytes[j] = bytes[j], bytes[i]
-	}
-}
-
-func bigIntFromLittleEndianBytes(bytes []byte) *big.Int {
-	reverse(bytes)
-	return new(big.Int).SetBytes(bytes)
-}
-
-func bigIntFromBigEndianBytes(bytes []byte) *big.Int {
-	return new(big.Int).SetBytes(bytes)
-}
-
-// size is number of bytes to return
-func littleEndianBytesFromBigInt(z *big.Int, size int) []byte {
-	bytes := z.Bytes()[0:size]
-	reverse(bytes)
-
-	return bytes
-}
-
-// size is number of bytes to return
-func bigEndianBytesFromBigInt(z *big.Int, size int) []byte {
-	return z.Bytes()[0:size]
-}
-
-func littleEndianBytesFromInt(n int, size int) []byte {
-	z := new(big.Int)
-	z.SetInt64(int64(n))
-	return littleEndianBytesFromBigInt(z, size)
-}
-
-func bigEndianBytesFromInt(n int, size int) []byte {
-	z := new(big.Int)
-	z.SetInt64(int64(n))
-	return bigEndianBytesFromBigInt(z, size)
-}
-
-func littleEndianBytesFromUInt(n UInt, size int) []byte {
-	z := new(big.Int)
-	z.SetUint64(uint64(n))
-	return littleEndianBytesFromBigInt(z, size)
-}
-
-func bigEndianBytesFromUInt(n UInt, size int) []byte {
-	z := new(big.Int)
-	z.SetUint64(uint64(n))
-	return bigEndianBytesFromBigInt(z, size)
-}
-
-func AsBytes_T(t util.T) []byte {
-	panic("Unimplemented for T")
-
-	return []byte{}
-}
-
-func AsBytes_UnsealedSectorCID(cid sector.UnsealedSectorCID) []byte {
-	panic("Unimplemented for UnsealedSectorCID")
-
-	return []byte{}
-}
-
-func AsBytes_SealedSectorCID(CID sector.SealedSectorCID) []byte {
-	panic("Unimplemented for SealedSectorCID")
-
-	return []byte{}
-}
-
-func AsBytes_PieceCID(CID piece.PieceCID) []byte {
-	panic("Unimplemented for PieceCID")
-
-	return []byte{}
-}
-
-func fromBytes_T(_ interface{}) util.T {
-	panic("Unimplemented for T")
-	return util.T{}
-}
-
-func isPow2(n int) bool {
-	return n != 0 && n&(n-1) == 0
 }
